Fixed a bunch of bugs. See description
*Fixed possible data leaks across characterSystem.h, and button.h
*Fixed utility.h not being able to be imported to multiple cpp files
*Moved inventorySystem.h to it's own cpp file
*Fixed drop item buttons not using the button.Draw function
*Fixed button.Draw function using legacy draw handles, moved to decals
*Added pragma once to button.h
*Added DrawRectDecal to olcPixelGameEngine.h
*Fixed programmer arithmetic error in items.cpp

+++ DeathByLiving/Scripts/turnManager.h
@@ -107,7 +107,7 @@ public:
 				}
 
 				// FIXME: draw movement/attack positions
-				for (size_t i = 0; i < 9; i++)
+				for (int i = 0; i < 9; i++)
 				{
 					if (i == 4 || (currentTurn->canMoveDiagonally == false && i % 2 == 0)) continue;
